requirements.txt
llama-index
streamlit
langchain
openai
pypdf
fastapi 
uvicorn 
prisma
jwt 
passlib
pinecone-client
google-search-results